requirements.txt
openai
langchain
langchain_openai
langchain_core
langchain_community
yfinance
pandas
pandas-ta
python-dotenv
colorama
scikit-learn>=1.0.0
xgboost>=1.5.0
numpy>=1.20.0
statsmodels>=0.13.0
matplotlib>=3.5.0
seaborn>=0.11.0
requests>=2.25.0
beautifulsoup4>=4.9.0
sentence-transformers>=2.2.0
aiohttp>=3.8.0
asyncio>=3.4.3
